requirements.txt
asgiref==3.2.10
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.3
backcall==0.1.0
bleach==3.1.1
cffi==1.14.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
Django==3.0.8
django-misaka==0.2.1
django-social-share==1.4.0
django-social-widgets==0.4.0
django-taggit==1.2.0
entrypoints==0.3
houdini.py==0.1.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
Markdown==3.2.1
MarkupSafe==1.1.1
mccabe==0.6.1
misaka==2.1.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.4
psycopg2-binary==2.8.4
pycodestyle==2.6.0
pycparser==2.20
Pygments==2.6.1
pylint==2.4.4
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2020.1
pywin32==227
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
Send2Trash==1.5.0
six==1.14.0
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
tornado==6.0.4
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
